requirements.txt
Django==1.11.4
django-extensions==1.8.1
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.1
pytz==2017.2
six==1.10.0
